sentinel/requirements.txt
# Core & Config
python-dotenv
pyyaml

# Backend: File System, Metadata, Content
filetype
Pillow
PyMuPDF
python-docx
pytesseract
psd-tools
osxmetadata; sys_platform == 'darwin' # macOS specific

# Backend: Database
sqlalchemy

# Backend: AI Inference
# Add llama-cpp-python later if needed

# UI
PyQt6 